+++ game_patch/misc/alpine_settings.cpp
@@ -1015,6 +1015,11 @@ bool alpine_player_settings_load(rf::Player* player)
         g_alpine_game_config.player_join_flash = std::stoi(settings["PlayerJoinFlash"]);
         processed_keys.insert("PlayerJoinFlash");
     }
+    if (settings.contains("PlayerJoinFlashTimeoutSeconds")) {
+        g_alpine_game_config.player_join_flash_timeout_sec =
+            std::stoul(settings["PlayerJoinFlashTimeoutSeconds"]);
+        processed_keys.insert("PlayerJoinFlashTimeoutSeconds");
+    }
     if (settings.contains("WorldHUDAltDamageIndicators")) {
         g_alpine_game_config.world_hud_alt_damage_indicators = std::stoi(settings["WorldHUDAltDamageIndicators"]);
         processed_keys.insert("WorldHUDAltDamageIndicators");
@@ -1485,6 +1490,7 @@ void alpine_player_settings_save(rf::Player* player)
     file << "DisableMultiCharacterLOD=" << g_alpine_game_config.multi_no_character_lod << "\n";
     file << "PlayerJoinBeep=" << g_alpine_game_config.player_join_beep << "\n";
     file << "PlayerJoinFlash=" << g_alpine_game_config.player_join_flash << "\n";
+    file << "PlayerJoinFlashTimeoutSeconds=" << g_alpine_game_config.player_join_flash_timeout_sec << "\n";
     file << "WorldHUDAltDamageIndicators=" << g_alpine_game_config.world_hud_alt_damage_indicators << "\n";
     file << "DesiredHandicap=" << g_alpine_game_config.desired_handicap << "\n";
     file << "CPOutlineHeightScale=" << g_alpine_game_config.control_point_outline_height_scale << "\n";