+++ game_patch/misc/alpine_settings.h
@@ -163,6 +163,7 @@ struct AlpineGameSettings
     int seasonal_effect = 1; // 0=none, 1=auto, 2=always_snow
     bool player_join_beep = false;
     bool player_join_flash = true;
+    uint32_t player_join_flash_timeout_sec = 0;
     bool full_range_lighting = true;
     bool always_clamp_official_lightmaps = false;
     bool ignore_tbl_vertex_lighting = false;